Project 0x11 C - printf

+++ _printf.c
@@ -48,19 +48,19 @@ int _printf(const char *format, ...)
 					count += print_int(va_arg(vl, int));
 					break;
 				case 'b':
-					count += print_bin(va_arg(vl, unsigned int));
+					count += print_bin(va_arg(vl, unsigned long int));
 					break;
 				case 'u':
-					count += print_unt(va_arg(vl, unsigned int));
+					count += print_unt(va_arg(vl, unsigned long int));
 					break;
 				case 'o':
-					count += print_oct(va_arg(vl, unsigned int));
+					count += print_oct(va_arg(vl, unsigned long int));
 					break;
 				case 'x':
-					count += print_hexlow(va_arg(vl, unsigned int));
+					count += print_hexlow(va_arg(vl, unsigned long int));
 					break;
 				case 'X':
-					count += print_hex(va_arg(vl, unsigned int));
+					count += print_hex(va_arg(vl, unsigned long int));
 					break;
 				case 'S':
 					count += print_Str(va_arg(vl, char *));
